+++ src/lib/stores.ts
@@ -1,13 +1,13 @@
 import { RuneStore } from "@tauri-store/svelte";
 import type { User } from "./graphql/twitch";
 import type { SidebarState } from "./hooks/use-sidebar.svelte";
-import type { SplitNode } from "./splits/types";
+import { LAYOUT_VERSION, type Layout } from "./splits/types";
 
 interface Storage {
 	[key: string]: unknown;
 	user: User | null;
 	accounts: User[];
-	layout: SplitNode | null;
+	layout: Layout | null;
 	pinned: string[];
 	sidebar: SidebarState;
 }
@@ -21,5 +21,16 @@ export const storage = new RuneStore<Storage>(
 		pinned: [],
 		sidebar: "expanded",
 	},
-	{ autoStart: true },
+	{
+		autoStart: true,
+		hooks: {
+			beforeFrontendSync: (state) => {
+				if (state.layout?.version !== LAYOUT_VERSION) {
+					state.layout = null;
+				}
+
+				return state;
+			},
+		},
+	},
 );
